+++ client/src/network/time_syncer.hpp
@@ -5,15 +5,15 @@
 
 class TimeSyncer {
   private:
-    float time_diff_to_server;
+    double time_diff_to_server;
     unsigned int minimum_responses = 10;
     unsigned int respones_handled = 0;
-    float minimum_latency_calculated = -1;
+    double minimum_latency_calculated = -1;
     bool finished_sync;
 
   public:
     std::vector<uint8_t> get_new_request();
     void handle_response(Msg::ServerToClient::TimeResponseData data);
-    float get_server_time();
+    double get_server_time();
     bool is_synchronized();
 };
